refactor image specific funcs to image package
